miscutil: provide getrandom() userland implementation if needed
Glibc 2.24 and older do not provide an interface for the getrandom()
system call. Linux kernels have been implementing this sytem call since
kernel 3.17.

Let's address the situation where glibc <= 2.24 and Linux >= 3.17 by
providing the system call interface ourselves when needed. This allows
users to take advantage of getrandom() as long as the kernel supports
it, even if glibc doesn't provide access to it.

Signed-off-by: Markus Mayer <mmayer@broadcom.com>